WEB SITE FOR YOUR PSYCHOLOGICAL COMFORT
This web site try to connects psychiatrists with patients. Any user/patient can use the site to: reserve an appointment with a doctor, submit a consultation to be answered by a doctor, read articles posted on the website that contains helpful information regards many topics of psychology a with chat bot responsible for appointments reservation process . Also the chat bot handles the consultation submitted by the user , All these under GDPR law to save information.